+++ Pods/Target Support Files/PathKit/PathKit-dummy.m
@@ -1,5 +1,0 @@
-#import <Foundation/Foundation.h>
-@interface PodsDummy_PathKit : NSObject
-@end
-@implementation PodsDummy_PathKit
-@end