+++ Sources/TestUtils/TestsHelper+Strings.swift
@@ -1,0 +1,55 @@
+//
+// SwiftGenKit UnitTests
+// Copyright © 2020 SwiftGen
+// MIT Licence
+//
+
+import Foundation
+import XCTest
+
+private func diff(_ result: String, _ expected: String) -> String? {
+  guard result != expected else { return nil }
+  var firstDiff: Int?
+  let newLines = CharacterSet.newlines
+  let lhsLines = result.components(separatedBy: newLines)
+  let rhsLines = expected.components(separatedBy: newLines)
+
+  for (idx, pair) in zip(lhsLines, rhsLines).enumerated() where pair.0 != pair.1 {
+    firstDiff = idx
+    break
+  }
+  if firstDiff == nil && lhsLines.count != rhsLines.count {
+    firstDiff = min(lhsLines.count, rhsLines.count)
+  }
+  if let badLineIdx = firstDiff {
+    let slice = { (lines: [String], context: Int) -> ArraySlice<String> in
+      let start = max(0, badLineIdx - context)
+      let end = min(badLineIdx + context, lines.count - 1)
+      return lines[start...end]
+    }
+    let addLineNumbers = { (slice: ArraySlice) -> [String] in
+      slice.enumerated().map { (idx: Int, line: String) in
+        let num = idx + slice.startIndex
+        let lineNum = "\(num + 1)".padding(toLength: 3, withPad: " ", startingAt: 0) + "|"
+        let clr = num == badLineIdx ? koCode : okCode
+        return "\(clr.num)\(lineNum)\(reset)\(clr.code)\(line)\(reset)"
+      }
+    }
+    let lhsNum = addLineNumbers(slice(lhsLines, 4)).joined(separator: "\n")
+    let rhsNum = addLineNumbers(slice(rhsLines, 4)).joined(separator: "\n")
+    return [
+      "\(msgColor)Mismatch at line \(badLineIdx)\(reset)",
+      ">>>>>> result",
+      "\(lhsNum)",
+      "======",
+      "\(rhsNum)",
+      "<<<<<< expected"
+    ].joined(separator: "\n")
+  }
+  return nil
+}
+
+public func XCTDiffStrings(_ result: String, _ expected: String, file: StaticString = #file, line: UInt = #line) {
+  guard let error = diff(result, expected) else { return }
+  XCTFail(error, file: file, line: line)
+}