+++ CHANGELOG.md
@@ -31,10 +31,27 @@ _None_
 * XCAssets: fixed some availability annotations that were incorrect.  
   [David Jennes](https://github.com/djbe)
   [#789](https://github.com/SwiftGen/SwiftGen/pull/789)
+* Strings: `objc-h` template now emits valid documentation comments.  
+  [@szotp](https://github.com/szotp)
+  [#822](https://github.com/SwiftGen/SwiftGen/pull/822)
+* Generate `xcfilelist`: Adds the template file path to the inputs `xcfilelist` (for custom templates).  
+  [Craig Siemens](https://github.com/CraigSiemens)
+  [#815](https://github.com/SwiftGen/SwiftGen/pull/815)
+* Strings: built-in templates now have better handling of multi-line translations.  
+  [@mrackwitz](https://github.com/mrackwitz)
+  [#774](https://github.com/SwiftGen/SwiftGen/pull/774)
 
 ### Internal Changes
 
-_None_
+* Switch to GitHub Actions.  
+  [#794](https://github.com/SwiftGen/SwiftGen/pull/794)
+  [David Jennes](https://github.com/djbe)
+* Switched the whole project over to use Swift Package Manager, restructuring some of the internals in the process.  
+  [David Jennes](https://github.com/djbe)
+  [#793](https://github.com/SwiftGen/SwiftGen/pull/793)
+* Updated Pods and Gems dependencies.  
+  [David Jennes](https://github.com/djbe)
+  [#827](https://github.com/SwiftGen/SwiftGen/pull/827)
 
 ## 6.4.0